Forgot to mention what I contributed
I made the game screen cover, win screen, and game over screen. Added code to connect the scenes to the game scene. I also helped with the tilemap and collected most of the assets needed. Mainly worked on the aesthetics and animations of the game, and a bit of coding for the player's key movements.